fapi/requirements.txt
# FastAPI 핵심
fastapi==0.109.0
uvicorn[standard]==0.27.0

# JWT 토큰 생성/검증
python-jose[cryptography]==3.3.0

# 비밀번호 해싱 (지금은 안 쓰지만 나중에 유용)
passlib[bcrypt]==1.7.4

# 파일 업로드 처리
python-multipart==0.0.6

# HTTP 요청 (구글 API 호출용)
requests==2.31.0

# 환경 변수 관리 (.env 파일 읽기)
python-dotenv==1.0.0

# 데이터베이스 ORM
sqlalchemy==2.0.25

# 세션 암호화 (SessionMiddleware용)
itsdangerous==2.1.2

# PostgreSQL 땜에
psycopg2==2.9.9